UI Polish: Inspector rounded corners + EntityInspector icons.
Inspector UI polish: rounded panel sections, entity type icons in header

Summary
Panel border radius — Added 8px border radius to all inspector sections (Hierarchy, Toolbar+Canvas, EntityInspector, SceneInfoPanel, Assets) by introducing a 4px padding and dark background on the .App container, making the rounded corners visible. Applied via the shared Box component class variants in App.css.

EntityHeader spacing — Increased the gap between the "+" (Add Component) button and the kebab (More Options) button from 4px to 8px in .EntityHeader .RightContent.

Entity type icons in EntityInspector — Extracted the HierarchyIcon logic from Hierarchy.tsx into a new shared EntityIcon component (Hierarchy/EntityIcon.tsx + EntityIcon.css). The icon is now shown next to the entity name in the EntityHeader, matching the icons already displayed in the Hierarchy panel. Cleaned up the now-redundant inline component and scoped CSS from Hierarchy.tsx/Hierarchy.css.

Test plan
 Open the inspector and verify all panels (Hierarchy, Canvas, EntityInspector, SceneInfoPanel, Assets) have visible 8px rounded corners with even black gaps between them
 Select an entity and verify its type icon appears in the EntityInspector header (entity, group, smart, custom, camera, player, tile)
 Add a Custom Item to the scene and verify only one icon appears in the header (no duplicates)
 Verify the Hierarchy still shows the correct icons for all entity types
