+++ packages/inspector/src/components/Hierarchy/Hierarchy.css
@@ -5,47 +5,6 @@
   padding: 6px;
 }
 
-.Hierarchy .tree-icon {
-  width: 16px;
-  height: 14px;
+.Hierarchy .entity-type-icon {
   margin-right: 4px;
-  background-size: 16px 16px;
-  background-position: center;
-  background-repeat: no-repeat;
-}
-
-.Hierarchy .player-icon {
-  background-image: url(./icons/player.svg);
-}
-
-.Hierarchy .camera-icon {
-  background-image: url(./icons/camera.svg);
-}
-
-.Hierarchy .custom-icon {
-  background-image: url(./icons/custom.svg);
-}
-
-.Hierarchy .smart-icon {
-  background-image: url(./icons/smart.svg);
-}
-
-.Hierarchy .tile-icon {
-  background-image: url(./icons/tile.svg);
-}
-
-.Hierarchy .entity-icon {
-  background-image: url(./icons/entity.svg);
-}
-
-.Hierarchy .group-icon {
-  background-image: url(./icons/group.svg);
-}
-
-.Hierarchy .spawn-area-icon {
-  background-image: url(./icons/spawn-area.svg);
-}
-
-.Hierarchy .camera-target-icon {
-  background-image: url(./icons/camera-target.svg);
 }
